Add support for sharding
I've added the ability to shard the top-level search space, as a cheap
way to parallelise across proceses. In-process multi-threaded
parallelisation would of course be quite doable, with a nicely
balanced work-queue and the like, but this is very cheap and easy, and
doesn't require me to pick up anything new.

The associated command-line parsing is implemented with clap, which
has been added as a Cargo dependency.